+++ package.json
@@ -20,6 +20,7 @@
     "react": "^17.0.2",
     "react-bootstrap": "^2.1.2",
     "react-dom": "^17.0.2",
+    "react-editor-md": "^1.0.31",
     "react-firebaseui": "^6.0.0",
     "react-markdown": "^8.0.0",
     "react-router-dom": "^6.2.1",
