+++ src/App.js
@@ -5,8 +5,8 @@ import LandingPage from "./pages/default/LandingPage";
 import AboutUs from "./pages/default/AboutUs";
 import Editor from "./pages/editors/Editor";
 import Courses from "./pages/Courses";
-import Community from "./pages/Community";
-import NotFound from "./pages/NotFound";
+import Community from "./pages/community/Community";
+import NotFound from "./pages/default/NotFound";
 import LogIn from "./pages/authentication/LogIn";
 import SignUp from "./pages/authentication/SignUp";
 import Confirmation from "./pages/authentication/Confirmation";
