+++ src/components/general/Footer.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Footer() {
+  return (
+    <div>
+        Footer
+    </div>
+  )
+}
+
+export default Footer
