--- src/pages/Community.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import Navbar from "../components/Navbar"
-
-const Community = () => {
-  return (
-    <div>
-        <Navbar/>
-    </div>
-  )
-}
-
-export default Community
