+++ src/pages/authentication/PasswordReset.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react'
-import Navbar from '../../components/Navbar'
+import Navbar from '../../components/general/Navbar'
 import { auth } from "../../firebase";
 import { sendPasswordResetEmail } from "firebase/auth"
