+++ src/pages/authentication/SignUp.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import Navbar from "../../components/Navbar";
+import Navbar from "../../components/general/Navbar";
 import { auth } from "../../firebase";
 import {
   createUserWithEmailAndPassword,
