+++ src/pages/community/Community.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import Navbar from "../../components/general/Navbar"
+import Footer from "../../components/general/Footer"
+
+const Community = () => {
+  return (
+    <div className='h-screen'>
+        <Navbar/>
+        <div className='h-full bg-gray-100'>
+
+        </div>
+    </div>
+  )
+}
+
+export default Community
