+++ src/pages/default/AboutUs.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
-import Markdown from "../../components/Markdown";
-import Navbar from '../../components/Navbar'
+import Markdown from "../../components/code-editor/Markdown";
+import Navbar from '../../components/general/Navbar'
 import FirebaseRequests from "../../components/FirebaseRequests";
 import StorageRequests from "../../components/StorageRequests";
