+++ src/pages/default/LandingPage.js
@@ -1,4 +1,4 @@
-import Navbar from '../../components/Navbar'
+import Navbar from '../../components/general/Navbar'
 
 function LandingPage() {
   return (
